requirements.txt
forcediphttpsadapter
homeassistant
numpy
PyYAML
requests
requests_cache
rdflib
websocket-client

voluptuous
PyJWT
aiohttp
astral
Jinja2
httpx
multidict
yarl
SQLAlchemy
attrs
urllib3
aiodns
bleak
python-dateutil
pillow
fnvhash
ifaddr
ciso8601
typing_extensions
httpcore
pytz
awesomeversion
packaging
bcrypt
certifi
orjson
python-slugify
cryptography
oauthlib
securetar
boto3
botocore
psutil
Flask